@@ -9,13 +9,10 @@ mainClassName = "hu.mondokm.ltlverif.MainProgram"
 group 'mondokm'
 version '1.0-SNAPSHOT'
 
-sourceCompatibility = 1.8
+sourceCompatibility = 1.17
 
 repositories {
     mavenCentral()
-    maven {
-        url  "https://dl.bintray.com/ftsrg/maven"
-    }
     flatDir {
         dirs 'lib'
     }
@@ -32,12 +29,25 @@ sourceSets {
 
 dependencies {
     antlr "org.antlr:antlr4:4.7.2"
-    compile group: 'theta', name: 'hu.bme.mit.theta.cfa', version: '0.0.1'
-    compile group: 'theta', name: 'hu.bme.mit.theta.core', version: '0.0.1'
-    compile group: 'theta', name: 'hu.bme.mit.theta.solver', version: '0.0.1'
-    compile group: 'theta', name: 'hu.bme.mit.theta.solver.z3', version: '0.0.1'
-    compile name: 'jhoafparser-1.1.1'
-    testCompile group: 'junit', name: 'junit', version: '4.12'
+    implementation group: 'theta', name: 'theta-analysis'
+    implementation group: 'theta', name: 'theta-cfa'
+    implementation group: 'theta', name: 'theta-cfa-analysis'
+    implementation group: 'theta', name: 'theta-common'
+    implementation group: 'theta', name: 'theta-core'
+    implementation group: 'theta', name: 'theta-solver'
+    implementation group: 'theta', name: 'theta-solver-z3'
+    implementation name: 'jhoafparser-1.1.1'
+    implementation name: 'owl-21.0'
+    implementation name: 'guava-31.1-jre'
+    testImplementation group: 'junit', name: 'junit', version: '4.12'
+}
+
+jar {
+    manifest {
+        attributes(
+                'Main-Class': 'hu.mondokm.ltlverif.MainProgram'
+        )
+    }
 }
 
 run {
