@@ -10,18 +10,21 @@ import hu.bme.mit.theta.solver.z3.Z3SolverFactory;
 import hu.mondokm.ltlverif.abstractor.ProductState;
 import hu.mondokm.ltlverif.buchi.BuchiAction;
 
+import java.util.Collection;
 import java.util.HashSet;
 
 public class CfaBuchiTransFunc {
 
-    private static PredTransFunc predTransFunc=PredTransFunc.create(PredAbstractors.booleanSplitAbstractor(Z3SolverFactory.getInstace().createSolver()));
+    private static PredTransFunc predTransFunc=PredTransFunc.create(PredAbstractors.booleanSplitAbstractor(Z3SolverFactory.getInstance().createSolver()));
 
     public static HashSet<ProductState> nextStates(CfaProductState curr, PredPrec precision){
         HashSet <ProductState> states=new HashSet<ProductState>();
         for(CFA.Edge edge: curr.getLoc().getOutEdges()) {
-            for(PredState state:predTransFunc.getSuccStates(curr.getPredState(), CfaAction.create(edge),precision)){
+            Collection<PredState> collection = predTransFunc.getSuccStates(curr.getPredState(), CfaAction.create(edge), precision);
+            for(PredState state : collection){
                 if(!state.isBottom())for(BuchiAction action: curr.getBuchiState().getActions()){
-                    for(PredState innerState:predTransFunc.getSuccStates(state,action,precision)){
+                    Collection<PredState> innerCollection = predTransFunc.getSuccStates(state,action,precision);
+                    for(PredState innerState : innerCollection){
                         if(!innerState.isBottom())states.add(new CfaProductState(edge, innerState, action.getTarget(),action));
                     }
                 }
